invoice penjualan #28

+++ application/controllers/Welcome.php
@@ -55,10 +55,15 @@ class Welcome extends CI_Controller {
 		$this->load->view('templates/footer');
 	}
 	public function proses_pesanan(){
-		$this->cart->destroy();
-		$this->load->view('templates/header');
-		$this->load->view('templates/sidebar');
-		$this->load->view('proses_pesanan');
-		$this->load->view('templates/footer');
+		$is_processed = $this->model_invoice->index();
+		if($is_processed){
+			$this->cart->destroy();
+			$this->load->view('templates/header');
+			$this->load->view('templates/sidebar');
+			$this->load->view('proses_pesanan');
+			$this->load->view('templates/footer');
+		}else{
+			echo "maaf, Pesanan Anda Gagal diproses!";
+		}
 	}
 }
